refactor: split StatisticsScreen into custom hook and modular components\n\n- Extracted useStatisticsData hook for data logic and caching\n- Created FrequencyList and ProgressBar components\n- Restored colored ball design for Lotto numbers\n- Improved maintainability, readability, and reusability
